extr_chem_info

This function takes a vector of IUPAC names and queries the PubChem database (using the webchem package) to obtain the corresponding CASRN and CID for each compound. It reshapes the resulting data, ensuring that each compound has a unique row with the CID, CASRN, and additional chemical properties.

Usage

extr_chem_info(IUPAC_names, stop_on_warning = FALSE)

Arguments

IUPAC_names

A character vector of IUPAC names. These are standardized names of chemical compounds that will be used to search in the PubChem database.

stop_on_warning

Logical. If set to TRUE, the function will stop and throw an error if any substances are not found in PubChem. Defaults to FALSE, in which case a warning is issued.

Value

A data frame with information on the queried compounds, including:

iupac_name

The IUPAC name of the compound.

cid

The PubChem Compound Identifier (CID).

isomeric_smiles

The SMILES string (Simplified Molecular Input Line Entry System).

Examples

if (FALSE) { # \dontrun{
# Example with formaldehyde and aflatoxin
extr_chem_info(IUPAC_names = c("Formaldehyde", "Aflatoxin B1"))
} # }