Extract Data from EPA IRIS Database

The extr_iris function sends a request to the EPA IRIS database to search for information based on a specified keywords and cancer types. It retrieves and parses the HTML content from the response. Note that if keywords is not provide all dataset are retrieved.

Usage

extr_iris(
  casrn = NULL,
  cancer_types = c("non_cancer", "cancer"),
  verbose = TRUE
)

Arguments

casrn

A vector CASRN for the search.

cancer_types

A character vector specifying the types of cancer to include in the search. Must be either "non_cancer" or "cancer".

verbose

A logical value indicating whether to print detailed messages. Default is TRUE.

Value

A data frame containing the extracted data.

See also

Examples

# \donttest{
extr_iris(c("1332-21-4", "50-00-0"))
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Quering 1332-21-4 to EPA IRIS database...
#> ℹ Request succeeded with status code: 200
#> ℹ Quering 50-00-0 to EPA IRIS database...
#> ℹ Request succeeded with status code: 200
#> # A tibble: 4 × 8
#>   chemical_name casrn     exposure_route assessment_type critical_effect_or_tu…¹
#>   <chr>         <chr>     <chr>          <chr>           <chr>                  
#> 1 Asbestos      1332-21-4 Inhalation     Cancer          Lung cancer and mesoth…
#> 2 Formaldehyde  50-00-0   Oral           Noncancer       Reduced weight gain, h…
#> 3 Formaldehyde  50-00-0   Inhalation     Cancer          Nasopharyngeal cancer,…
#> 4 Formaldehyde  50-00-0   Inhalation     Noncancer       Decreased pulmonary fu…
#> # ℹ abbreviated name: ¹​critical_effect_or_tumor_type
#> # ℹ 3 more variables: woe_characterization <chr>, toxicity_value_type <chr>,
#> #   toxicity_value <chr>
# }