Retrieve WHO IARC Monograph Information

This function retrieves information regarding Monographs from the World Health Organization (WHO) International Agency for Research on Cancer (IARC) based on CAS Registry Number or Name of the chemical.

Usage

extr_monograph(ids, search_type = "casrn", verbose = TRUE)

Arguments

ids

A character vector of IDs to search for.

search_type

A character string specifying the type of search to perform. Valid options are "cas_rn" (CAS Registry Number) and "name" (name of the chemical). If search_type is "casrn", the function filters by the CAS Registry Number. . If search_type is "name", the function performs a partial match search for the chemical name.

verbose

A logical value indicating whether to print detailed messages. Default is TRUE.

Value

A data frame containing the relevant information from the WHO IARC, including Monograph volume, volume_publication_year, evaluation_year, and additional_information where the chemical was described.

See also

Examples

{
  extr_monograph(search_type = "casrn", ids = c("105-74-8", "120-58-1"))

  # Example usage for name search
  extr_monograph(search_type = "name", ids = c("Aloe", "Schistosoma", "Styrene"))
}
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Extracting WHO IARC monographs...
#> Last updated: 2024-11-29 5:08pm (CET)
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Extracting WHO IARC monographs...
#> Last updated: 2024-11-29 5:08pm (CET)
#>          casrn                                    agent group         volume
#> 1                         Aloe vera, whole leaf extract    2B            108
#> 9              Schistosoma haematobium (infection with)     1       61, 100B
#> 10               Schistosoma japonicum (infection with)    2B             61
#> 11                 Schistosoma mansoni (infection with)     3             61
#> 220   100-42-5                                  Styrene    2A    60, 82, 121
#> 1045 9003-54-7         Styrene-acrylonitrile copolymers     3      19, Sup 7
#> 1046 9003-55-8             Styrene-butadiene copolymers     3      19, Sup 7
#> 1087   96-09-3                        Styrene-7,8-oxide    2A Sup 7, 60, 121
#>      volume_publication_year evaluation_year
#> 1                       2016            2013
#> 9                       2012            2009
#> 10                      1994            1994
#> 11                      1994            1994
#> 220                     2019            2018
#> 1045                    1987            1987
#> 1046                    1987            1987
#> 1087                    2019            2018
#>                                                                            additional_information
#> 1                                                                                                
#> 9                                                                                                
#> 10                                                                                               
#> 11                                                                                               
#> 220                                                                                              
#> 1045                                                                                             
#> 1046                                                                                             
#> 1087 NB Overall evaluation upgraded to Group 2A with supporting evidence from other relevant data