Extract Data from EPA PPRTVs

Extracts data for specified identifiers (CASRN or chemical names) from the EPA's Provisional Peer-Reviewed Toxicity Values (PPRTVs) database. The function retrieves and processes data, with options to use cached files or force a fresh download.

Usage

extr_pprtv(
  ids,
  search_type = "casrn",
  verbose = TRUE,
  force = TRUE,
  get_all = FALSE
)

Arguments

ids

Character vector of identifiers to search (e.g., CASRN or chemical names).

search_type

Character string specifying the type of identifier: "casrn" or "name". Default is "casrn". If search_type is "name", the function performs a partial match search for the chemical name. NOTE: Since partial mached is use, multiple seraches might match the same chemical, therefore chemical ids might not be uniques.

verbose

Logical indicating whether to display progress messages. Default is TRUE.

force

Logical indicating whether to force a fresh download of the database. Default is TRUE.

get_all

Logical. If TRUE ignore all the other ignore ids, search_type, set force = TRUE and get the all dataset. This is was introduced for debugging purposes.

Value

A data frame with extracted information matching the specified identifiers, or NULL if no matches are found.

See also

Examples

# \donttest{
condathis::with_sandbox_dir({ # this is to write on tempdir as for CRAN policies # nolint
  # Extract data for a specific CASRN
  extr_pprtv(ids = "107-02-8", search_type = "casrn", verbose = TRUE)

  # Extract data for a chemical name
  out <- extr_pprtv(
    ids = "Acrolein", search_type = "name", verbose = TRUE,
    force = FALSE
  )
  print(out)

  # Extract data for multiple identifiers
  out2 <- extr_pprtv(
    ids = c("107-02-8", "79-10-7", "42576-02-3"),
    search_type = "casrn",
    verbose = TRUE,
    force = FALSE
  )
  print(out2)
})
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Downloading data from <https://cfpub.epa.gov/ncea/pprtv/atoz.cfm>.
#> ✔ micromamba successfully downloaded.
#> Running \
#>   /private/var/folders/bp/fpwcfq1563l21rz5gdcsfcsw0000gn/T/RtmpFBC2H8/tmp-datacc6514dfce72/R/condathis/micromamba/bin/micromamba \
#>   --no-rc --no-env create -r \
#>   /private/var/folders/bp/fpwcfq1563l21rz5gdcsfcsw0000gn/T/RtmpFBC2H8/tmp-datacc6514dfce72/R/condathis \
#>   -n openssl-linux-env --yes --no-channel-priority --override-channels \
#>   '--channel-priority=0' -c bioconda -c conda-forge 'curl==7.78.0' libcurl \
#>   openssl
#> 
#> 
#> Transaction
#> 
#>   Prefix: /private/var/folders/bp/fpwcfq1563l21rz5gdcsfcsw0000gn/T/RtmpFBC2H8/tmp-datacc6514dfce72/R/condathis/envs/openssl-linux-env
#> 
#>   Updating specs:
#> 
#>    - curl==7.78.0
#>    - libcurl
#>    - openssl
#> 
#> 
#>   Package                 Version  Build             Channel          Size
#> ────────────────────────────────────────────────────────────────────────────
#>   Install:
#> ────────────────────────────────────────────────────────────────────────────
#> 
#>   + c-ares                 1.34.5  hf13058a_0        conda-forge     185kB
#>   + ca-certificates     2025.4.26  hbd8a1cb_0        conda-forge     152kB
#>   + curl                   7.78.0  hb861fe1_0        conda-forge     140kB
#>   + krb5                   1.19.3  hb49756b_0        conda-forge       1MB
#>   + libcurl                7.78.0  hf45b732_0        conda-forge     324kB
#>   + libcxx                 20.1.5  hf95d169_0        conda-forge     562kB
#>   + libedit          3.1.20250104  pl5321ha958ccf_0  conda-forge     116kB
#>   + libev                    4.33  h10d778d_2        conda-forge     107kB
#>   + libnghttp2             1.51.0  h0dd9d14_0        conda-forge     609kB
#>   + libssh2                1.10.0  h7535e13_3        conda-forge     227kB
#>   + libzlib                1.2.13  h87427d6_6        conda-forge      57kB
#>   + ncurses                   6.5  h0622a9a_3        conda-forge     822kB
#>   + openssl                1.1.1w  h8a1eda9_0        conda-forge       2MB
#>   + zlib                   1.2.13  h87427d6_6        conda-forge      89kB
#> 
#>   Summary:
#> 
#>   Install: 14 packages
#> 
#>   Total download: 6MB
#> 
#> ────────────────────────────────────────────────────────────────────────────
#> 
#> 
#> 
#> Transaction starting
#> Linking libzlib-1.2.13-h87427d6_6
#> Linking libcxx-20.1.5-hf95d169_0
#> Linking c-ares-1.34.5-hf13058a_0
#> Linking libev-4.33-h10d778d_2
#> Linking ncurses-6.5-h0622a9a_3
#> Linking zlib-1.2.13-h87427d6_6
#> Linking libedit-3.1.20250104-pl5321ha958ccf_0
#> Linking ca-certificates-2025.4.26-hbd8a1cb_0
#> Linking openssl-1.1.1w-h8a1eda9_0
#> Linking krb5-1.19.3-hb49756b_0
#> Linking libssh2-1.10.0-h7535e13_3
#> Linking libnghttp2-1.51.0-h0dd9d14_0
#> Linking libcurl-7.78.0-hf45b732_0
#> Linking curl-7.78.0-hb861fe1_0
#> 
#> Transaction finished
#> 
#> To activate this environment, use:
#> 
#>     micromamba activate openssl-linux-env
#> 
#> Or to execute a single command in this environment, use:
#> 
#>     micromamba run -n openssl-linux-env mycommand
#> 
#> ℹ Saving data in the cache /private/var/folders/bp/fpwcfq1563l21rz5gdcsfcsw0000gn/T/RtmpFBC2H8/tmp-homecc65556db277/Library/Caches/org.R-project.R/R/extractox/epa_pprtvs.rds.
#> ℹ Extracting EPA PPRTVs.
#> ✔ Successfully load. epa_pprtvs.rds from cache.
#> ℹ Cache date Mon, 26 May 2025 18:36:49 GMT.
#> ℹ Set `force = TRUE` to force download from EPA.
#> ℹ Extracting EPA PPRTVs.
#>   pprtv_substance_id chemical    casrn last_revision
#> 1               1555 Acrolein 107-02-8          2002
#>                                          pprtv_assessment
#> 1 https://cfpub.epa.gov/ncea/pprtv/documents/Acrolein.pdf
#>                                                                 iris_link
#> 1 https://cfpub.epa.gov/ncea/iris2/chemicalLanding.cfm?substance_nmbr=364
#>   rf_c_value rf_d_value           woe               date_downloaded    query
#> 1   See IRIS   See IRIS Not available Mon, 26 May 2025 18:36:49 GMT Acrolein
#> ✔ Successfully load. epa_pprtvs.rds from cache.
#> ℹ Cache date Mon, 26 May 2025 18:36:49 GMT.
#> ℹ Set `force = TRUE` to force download from EPA.
#> ℹ Extracting EPA PPRTVs.
#>   pprtv_substance_id     chemical      casrn last_revision
#> 1               1555     Acrolein   107-02-8          2002
#> 2               1594      Bifenox 42576-02-3          2006
#> 3               1556 Acrylic Acid    79-10-7          2010
#>                                             pprtv_assessment
#> 1    https://cfpub.epa.gov/ncea/pprtv/documents/Acrolein.pdf
#> 2     https://cfpub.epa.gov/ncea/pprtv/documents/Bifenox.pdf
#> 3 https://cfpub.epa.gov/ncea/pprtv/documents/AcrylicAcid.pdf
#>                                                                 iris_link
#> 1 https://cfpub.epa.gov/ncea/iris2/chemicalLanding.cfm?substance_nmbr=364
#> 2                                                           Not available
#> 3   https://cfpub.epa.gov/ncea/iris2/chemicalLanding.cfm?substance_nmbr=2
#>      rf_c_value         rf_d_value
#> 1      See IRIS           See IRIS
#> 2 Not available 9 x 10-3 mg/kg-day
#> 3      See IRIS           See IRIS
#>                                                       woe
#> 1                                           Not available
#> 2 Inadequate information to assess carcinogenic potential
#> 3 Inadequate information to assess carcinogenic potential
#>                 date_downloaded      query
#> 1 Mon, 26 May 2025 18:36:49 GMT   107-02-8
#> 2 Mon, 26 May 2025 18:36:49 GMT 42576-02-3
#> 3 Mon, 26 May 2025 18:36:49 GMT    79-10-7
# }