Extract Toxicological Information from Multiple Databases

This wrapper function retrieves toxicological information for specified chemicals by calling several external functions to query multiple databases, including PubChem, the Integrated Chemical Environment (ICE), CompTox Chemicals Dashboard, and the Integrated Risk Information System (IRIS).

Usage

extr_tox(casrn)

Arguments

casrn

A character vector of CAS Registry Numbers (CASRN) representing the chemicals of interest.

Value

A list of data frames containing toxicological information retrieved from each database:

ghs_dat

Toxicity data from PubChem's Globally Harmonized System (GHS) classification.

comptox_list

List of toxicity information from the CompTox Chemicals Dashboard.

ice_dat

Assay data from the Integrated Chemical Environment (ICE) database.

iris

Risk assessment data from the IRIS database.

Details

Specifically, this function:

Calls extr_pubchem_ghs to retrieve GHS classification data from PubChem.

Calls extr_comptox to retrieve data from the CompTox Chemicals Dashboard.

Calls extr_ice to gather assay data from the ICE database.

Calls extr_iris to retrieve risk assessment information from the IRIS database.

Examples

# \donttest{
casrn <- c("50-00-0", "107-02-8")
extr_tox(casrn)
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Getting PubChem IDS...
#> Querying 50-00-0. 
#> OK (HTTP 200).
#> 
#> 
#> Querying 712. 
#> OK (HTTP 200).
#> 
#> 
#> ℹ Getting PubChem IDS...
#> Querying 107-02-8. 
#> OK (HTTP 200).
#> 
#> 
#> Querying 7847. 
#> OK (HTTP 200).
#> 
#> 
#> ℹ Sending request to CompTox...
#> Error in extr_comptox_(ids = ids, download_items = download_items, mass_error = mass_error,     verify_ssl = verify_ssl, xlsx_file = xlsx_file, base_url = base_url,     ...): Failed to perform the request: Failed to perform HTTP request. Caused by
#> error in `curl::curl_fetch_memory()`: ! SSL connect error [comptox.epa.gov]:
#> error:0A000152:SSL routines::unsafe legacy renegotiation disabled
#> 
#> !If you are using openssl, you might need to downgrade to curl v7.78.0, openssl
#> v1.1.1!
# }