Extract Data from EPA PPRTVs

Extracts data for specified identifiers (CASRN or chemical names) from the EPA's Provisional Peer-Reviewed Toxicity Values (PPRTVs) database. The function retrieves and processes data, with options to use cached files or force a fresh download.

Usage

extr_pprtv(
  ids,
  search_type = "casrn",
  verbose = TRUE,
  force = TRUE,
  get_all = FALSE
)

Arguments

ids

Character vector of identifiers to search (e.g., CASRN or chemical names).

search_type

Character string specifying the type of identifier: "casrn" or "name". Default is "casrn". If search_type is "name", the function performs a partial match search for the chemical name. NOTE: Since partial mached is use, multiple seraches might match the same chemical, therefore chemical ids might not be uniques.

verbose

Logical indicating whether to display progress messages. Default is TRUE.

force

Logical indicating whether to force a fresh download of the database. Default is TRUE.

get_all

Logical. If TRUE ignore all the other ignore ids, search_type, set force = TRUE and get the all dataset. This is was introduced for debugging purposes.

Value

A data frame with extracted information matching the specified identifiers, or NULL if no matches are found.

See also

Examples

# \donttest{
condathis::with_sandbox_dir({ # this is to write on tempdir as for CRAN policies # nolint
  # Extract data for a specific CASRN
  extr_pprtv(ids = "107-02-8", search_type = "casrn", verbose = TRUE)

  # Extract data for a chemical name
  out <- extr_pprtv(
    ids = "Acrolein", search_type = "name", verbose = TRUE,
    force = FALSE
  )
  print(out)

  # Extract data for multiple identifiers
  out2 <- extr_pprtv(
    ids = c("107-02-8", "79-10-7", "42576-02-3"),
    search_type = "casrn",
    verbose = TRUE,
    force = FALSE
  )
  print(out2)
})
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Downloading data from <https://cfpub.epa.gov/ncea/pprtv/atoz.cfm>.
#> ✔ micromamba successfully downloaded.
#> Running \
#>   /tmp/RtmpYdMnNJ/tmp-data21a9fbd048e4e/R/condathis/micromamba/bin/micromamba \
#>   --no-rc --no-env create -r \
#>   /tmp/RtmpYdMnNJ/tmp-data21a9fbd048e4e/R/condathis -n openssl-linux-env \
#>   --yes --no-channel-priority --override-channels '--channel-priority=0' -c \
#>   bioconda -c conda-forge 'curl==7.78.0' libcurl openssl
#> 
#> 
#> Transaction
#> 
#>   Prefix: /tmp/RtmpYdMnNJ/tmp-data21a9fbd048e4e/R/condathis/envs/openssl-linux-env
#> 
#>   Updating specs:
#> 
#>    - curl==7.78.0
#>    - libcurl
#>    - openssl
#> 
#> 
#>   Package                 Version  Build             Channel          Size
#> ────────────────────────────────────────────────────────────────────────────
#>   Install:
#> ────────────────────────────────────────────────────────────────────────────
#> 
#>   + _libgcc_mutex             0.1  conda_forge       conda-forge       3kB
#>   + _openmp_mutex             4.5  2_gnu             conda-forge      24kB
#>   + c-ares                 1.34.5  hb9d3cd8_0        conda-forge     207kB
#>   + ca-certificates     2025.6.15  hbd8a1cb_0        conda-forge     151kB
#>   + curl                   7.78.0  hea6ffbf_0        conda-forge     151kB
#>   + keyutils                1.6.1  h166bdaf_0        conda-forge     118kB
#>   + krb5                   1.19.3  h3790be6_0        conda-forge       2MB
#>   + libcurl                7.78.0  h2574ce0_0        conda-forge     343kB
#>   + libedit          3.1.20250104  pl5321h7949ede_0  conda-forge     135kB
#>   + libev                    4.33  hd590300_2        conda-forge     113kB
#>   + libgcc                 15.1.0  h767d61c_3        conda-forge     825kB
#>   + libgcc-ng              15.1.0  h69a702a_3        conda-forge      29kB
#>   + libgomp                15.1.0  h767d61c_3        conda-forge     447kB
#>   + libnghttp2             1.51.0  hdcd2b5c_0        conda-forge     623kB
#>   + libssh2                1.10.0  haa6b8db_3        conda-forge     239kB
#>   + libstdcxx              15.1.0  h8f9b012_3        conda-forge       4MB
#>   + libstdcxx-ng           15.1.0  h4852527_3        conda-forge      29kB
#>   + libzlib                1.2.13  h4ab18f5_6        conda-forge      62kB
#>   + ncurses                   6.5  h2d0b736_3        conda-forge     892kB
#>   + openssl                1.1.1w  hd590300_0        conda-forge       2MB
#>   + zlib                   1.2.13  h4ab18f5_6        conda-forge      93kB
#> 
#>   Summary:
#> 
#>   Install: 21 packages
#> 
#>   Total download: 12MB
#> 
#> ────────────────────────────────────────────────────────────────────────────
#> 
#> 
#> 
#> Transaction starting
#> Linking ca-certificates-2025.6.15-hbd8a1cb_0
#> Linking libgomp-15.1.0-h767d61c_3
#> Linking _libgcc_mutex-0.1-conda_forge
#> Linking _openmp_mutex-4.5-2_gnu
#> Linking libgcc-15.1.0-h767d61c_3
#> Linking ncurses-6.5-h2d0b736_3
#> Linking c-ares-1.34.5-hb9d3cd8_0
#> Linking libstdcxx-15.1.0-h8f9b012_3
#> Linking libgcc-ng-15.1.0-h69a702a_3
#> Linking libedit-3.1.20250104-pl5321h7949ede_0
#> Linking libstdcxx-ng-15.1.0-h4852527_3
#> Linking libev-4.33-hd590300_2
#> Linking keyutils-1.6.1-h166bdaf_0
#> Linking libzlib-1.2.13-h4ab18f5_6
#> Linking openssl-1.1.1w-hd590300_0
#> Linking zlib-1.2.13-h4ab18f5_6
#> Linking krb5-1.19.3-h3790be6_0
#> Linking libssh2-1.10.0-haa6b8db_3
#> Linking libnghttp2-1.51.0-hdcd2b5c_0
#> Linking libcurl-7.78.0-h2574ce0_0
#> Linking curl-7.78.0-hea6ffbf_0
#> 
#> Transaction finished
#> 
#> To activate this environment, use:
#> 
#>     micromamba activate openssl-linux-env
#> 
#> Or to execute a single command in this environment, use:
#> 
#>     micromamba run -n openssl-linux-env mycommand
#> 
#> ℹ Saving data in the cache /tmp/RtmpYdMnNJ/tmp-home21a9fb2c6238fc/.cache/R/extractox/epa_pprtvs.rds.
#> ℹ Extracting EPA PPRTVs.
#> ✔ Successfully load. epa_pprtvs.rds from cache.
#> ℹ Cache date Tue, 08 Jul 2025 19:48:03 GMT.
#> ℹ Set `force = TRUE` to force download from EPA.
#> ℹ Extracting EPA PPRTVs.
#>   pprtv_substance_id chemical    casrn last_revision
#> 1               1555 Acrolein 107-02-8          2002
#>                                          pprtv_assessment
#> 1 https://cfpub.epa.gov/ncea/pprtv/documents/Acrolein.pdf
#>                                                                 iris_link
#> 1 https://cfpub.epa.gov/ncea/iris2/chemicalLanding.cfm?substance_nmbr=364
#>   rf_c_value rf_d_value           woe               date_downloaded    query
#> 1   See IRIS   See IRIS Not available Tue, 08 Jul 2025 19:48:03 GMT Acrolein
#> ✔ Successfully load. epa_pprtvs.rds from cache.
#> ℹ Cache date Tue, 08 Jul 2025 19:48:03 GMT.
#> ℹ Set `force = TRUE` to force download from EPA.
#> ℹ Extracting EPA PPRTVs.
#>   pprtv_substance_id     chemical      casrn last_revision
#> 1               1555     Acrolein   107-02-8          2002
#> 2               1594      Bifenox 42576-02-3          2006
#> 3               1556 Acrylic Acid    79-10-7          2010
#>                                             pprtv_assessment
#> 1    https://cfpub.epa.gov/ncea/pprtv/documents/Acrolein.pdf
#> 2     https://cfpub.epa.gov/ncea/pprtv/documents/Bifenox.pdf
#> 3 https://cfpub.epa.gov/ncea/pprtv/documents/AcrylicAcid.pdf
#>                                                                 iris_link
#> 1 https://cfpub.epa.gov/ncea/iris2/chemicalLanding.cfm?substance_nmbr=364
#> 2                                                           Not available
#> 3   https://cfpub.epa.gov/ncea/iris2/chemicalLanding.cfm?substance_nmbr=2
#>      rf_c_value         rf_d_value
#> 1      See IRIS           See IRIS
#> 2 Not available 9 x 10-3 mg/kg-day
#> 3      See IRIS           See IRIS
#>                                                       woe
#> 1                                           Not available
#> 2 Inadequate information to assess carcinogenic potential
#> 3 Inadequate information to assess carcinogenic potential
#>                 date_downloaded      query
#> 1 Tue, 08 Jul 2025 19:48:03 GMT   107-02-8
#> 2 Tue, 08 Jul 2025 19:48:03 GMT 42576-02-3
#> 3 Tue, 08 Jul 2025 19:48:03 GMT    79-10-7
# }