Write Dataframes to Excel

This function creates an Excel file with each dataframe in a list as a separate sheet.

Usage

write_dataframes_to_excel(df_list, filename)

Arguments

df_list

A named list of dataframes to write to the Excel file.

filename

The name of the Excel file to create.

Value

No return value. The function prints a message indicating the completion of the Excel file writing.

Examples

# \donttest{
tox_dat <- extr_tox("50-00-0")
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Getting PubChem IDS...
#> Querying 50-00-0. 
#> OK (HTTP 200).
#> 
#> Querying 712. 
#> OK (HTTP 200).
#> 
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Sending request to CompTox...
#> ℹ Request succeeded with status code: 202
#> ℹ Getting info from CompTox...
#> ℹ Request succeeded with status code: 200
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Sending request to ICE database...
#> ℹ Request succeeded with status code: 200
#> ℹ Checking Internet Connection...
#> ℹ Internet connection OK...
#> ℹ Quering 50-00-0 to EPA IRIS database...
#> ℹ Request succeeded with status code: 200
#> ℹ Extracting WHO IARC monographs...
#> Last updated: 2024-11-29 5:08pm (CET)
temp_file <- tempfile(fileext = ".xlsx")
write_dataframes_to_excel(tox_dat, filename = temp_file)
#> ℹ Excell file written in /var/folders/bp/fpwcfq1563l21rz5gdcsfcsw0000gn/T//RtmplDlkBc/file17c226ffaed9d.xlsx...
# }